+++ xskillscore/tests/test_accessor_probabilistic.py
@@ -11,6 +11,7 @@ from xskillscore.core.probabilistic import (
     crps_quadrature,
     discrimination,
     rank_histogram,
+    reliability,
     rps,
     threshold_brier_score,
 )
@@ -108,36 +109,45 @@ def test_rank_histogram_accessor(o, f_prob, outer_bool):
 
 @pytest.mark.parametrize('outer_bool', [False, True])
 def test_discrimination_accessor(o, f_prob, threshold, outer_bool):
-    hist_event_actual, hist_no_event_actual = discrimination(
-        o > threshold, (f_prob > threshold).mean('member')
-    )
+    actual = discrimination(o > threshold, (f_prob > threshold).mean('member'))
     ds = xr.Dataset()
     ds['o'] = o > threshold
     ds['f_prob'] = (f_prob > threshold).mean('member')
     if outer_bool:
         ds = ds.drop_vars('f_prob')
-        hist_event_expected, hist_no_event_expected = ds.xs.discrimination(
-            'o', (f_prob > threshold).mean('member')
-        )
+        expected = ds.xs.discrimination('o', (f_prob > threshold).mean('member'))
     else:
-        hist_event_expected, hist_no_event_expected = ds.xs.discrimination(
-            'o', 'f_prob'
-        )
-    assert_allclose(hist_event_actual, hist_event_expected)
-    assert_allclose(hist_no_event_actual, hist_no_event_expected)
+
+        expected = ds.xs.discrimination('o', 'f_prob')
+    assert_allclose(actual, expected)
+
+
+@pytest.mark.parametrize('outer_bool', [False, True])
+def test_reliability_accessor(o, f_prob, threshold, outer_bool):
+    actual = reliability(o > threshold, (f_prob > threshold).mean('member'))
+
+    ds = xr.Dataset()
+    ds['o'] = o > threshold
+    ds['f_prob'] = (f_prob > threshold).mean('member')
+    if outer_bool:
+        ds = ds.drop_vars('f_prob')
+        expected = ds.xs.reliability('o', (f_prob > threshold).mean('member'))
+    else:
+        expected = ds.xs.reliability('o', 'f_prob')
+    assert_allclose(actual, expected)
 
 
 @pytest.mark.parametrize('outer_bool', [False, True])
 def test_rps_accessor(o, f_prob, outer_bool):
     category_edges = np.linspace(0, 1 + 1e-8, 6)
-    expected = rps(o, f_prob, category_edges=category_edges)
+    actual = rps(o, f_prob, category_edges=category_edges)
 
     ds = xr.Dataset()
     ds['o'] = o
     ds['f_prob'] = f_prob
     if outer_bool:
         ds = ds.drop_vars('f_prob')
-        actual = ds.xs.rps('o', f_prob, category_edges=category_edges)
+        expected = ds.xs.rps('o', f_prob, category_edges=category_edges)
     else:
-        actual = ds.xs.rps('o', 'f_prob', category_edges=category_edges)
+        expected = ds.xs.rps('o', 'f_prob', category_edges=category_edges)
     assert_allclose(actual, expected)