+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Features
 - Added contingency table and associated metrics (:pr:`119`, :pr:`153`). `Dougie Squire`_
 - Added ``rank_histogram`` and ``discrimination`` to probabilistic metrics (:pr:`136`).
   `Dougie Squire`_
+- Added ``reliability`` to probabilistic metrics (:pr:`164`). `Dougie Squire`_
 - Added ``dim`` and ``weights`` for probabilistic metrics. (:pr:`121`) `Aaron Spring`_
 - Added ``keep_attrs`` for all metrics. (:pr:`122`) `Andrew Huang`_
 - Added ranked probability score ``rps``. (:pr:`163`) `Aaron Spring`_